# Profile/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для управления профилями пользователей.",
        "title": "Profile Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/profile/api/v1",
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Получить свой профиль",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile"
                        }
                    },
                    "204": {
                        "description": "Профиль не активен"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Полное обновление профиля",
                "parameters": [
                    {
                        "description": "Новые данные профиля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.ProfileInfoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile"
                        }
                    },
                    "409": {
                        "description": "Conflict - Field already taken",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает профиль пользователя с соцсетями и целями",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Создать новый профиль",
                "parameters": [
                    {
                        "description": "Данные профиля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.CreateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Профиль успешно создан",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - Field already taken",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Мягкое удаление своего профиля",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Частично обновить профиль",
                "parameters": [
                    {
                        "description": "Поля для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UpdateProfilePartialDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile"
                        }
                    }
                }
            }
        },
        "/profile/purpose/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "purpose"
                ],
                "summary": "Удалить цель",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID цели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/profile/purposes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purpose"
                ],
                "summary": "Создать новые цели",
                "parameters": [
                    {
                        "description": "Список целей",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UserPurposeDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UserPurposeDTO"
                            }
                        }
                    }
                }
            }
        },
        "/profile/social": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Добавить ссылки на соцсети",
                "parameters": [
                    {
                        "description": "Список ссылок",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.SocialLinkDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.SocialLinkDTO"
                            }
                        }
                    }
                }
            }
        },
        "/profile/social/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Обновить конкретную соц. ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ссылки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные ссылки",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.SocialLinkDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.SocialLinkDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not your link",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "social"
                ],
                "summary": "Удалить соц. ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ссылки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/profile/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет просмотреть чужой профиль по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Получить профиль по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID профиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Полное удаление профиля (Admin only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/profiles/purpose/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "purpose"
                ],
                "summary": "Обновить конкретную цель",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID цели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные цели",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UserPurposeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UserPurposeDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_YoungFlores_Case_Go_Profile_internal_profile_dto.UserPurposeDTO": {
            "type": "object",
            "required": [
                "purpose"
            ],
            "properties": {
                "purpose": {
                    "type": "string",
                    "minLength": 5
                }
            }
        },
        "github_com_YoungFlores_Case_Go_Profile_internal_profile_models.Profile": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "case_count": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/github_com_YoungFlores_Case_Go_Profile_internal_profile_models.UserSex"
                },
                "surname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}